Adding @SpringBootTest using TestRestTemplate writing two test which test if page is loading (/car-dealership and /mechanic)

+++ src/main/resources/db/migration/V1_4__car_dealership_car_to_buy.sql
@@ -4,7 +4,7 @@ CREATE TABLE car_to_buy
     vin             VARCHAR(20)    NOT NULL,
     brand           VARCHAR(32)    NOT NULL,
     model           VARCHAR(32)    NOT NULL,
-    production_year SMALLINT       NOT NULL,
+    production_year INT            NOT NULL,
     color           VARCHAR(32),
     price           NUMERIC(19, 2) NOT NULL,
     PRIMARY KEY (car_to_buy_id),
