+++ src/main/resources/db/migration/V1_5__car_dealership_car_to_service.sql
@@ -4,7 +4,7 @@ CREATE TABLE car_to_service
     vin               VARCHAR(20) NOT NULL,
     brand             VARCHAR(32) NOT NULL,
     model             VARCHAR(32) NOT NULL,
-    production_year   SMALLINT    NOT NULL,
+    production_year   INT         NOT NULL,
     PRIMARY KEY (car_to_service_id),
     UNIQUE (vin)
 );
